Update VIKOR.R
if( (Q == "NaN") || (Q == "Inf")) 
It used to work before with a warning (R ver. 4.2.0). But in R version 4.3.0, it returns an error
Calling && or || with LHS or (if evaluated) RHS of length greater than one is now always an error, with a report  of the form
'length = 4' in coercion to 'logical(1)'